dev-requirements.txt
# Test dependencies
pytest==3.0.7
pytest-django==3.1.2